Split gcat and its underlying block behaviors
